add gilded rose controller test
