+++ src/main/java/com/example/gildedrose/Application.java
@@ -1,4 +1,4 @@
-package com.example.springboot;
+package com.example.gildedrose;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
