+++ src/main/java/com/example/gildedrose/item/Itemable.java
@@ -1,4 +1,4 @@
-package gildedroseinventory;
+package com.example.gildedrose.item;
 
 public interface Itemable {
     public void update();
