+++ src/main/java/com/example/gildedrose/item/NormalItem.java
@@ -1,22 +1,23 @@
-package gildedroseinventory;
+package com.example.gildedrose.item;
 
-public class AgedBrieItem implements Itemable{
+public class NormalItem implements Itemable{
     public final String name;
     public int sellIn;
     public int quality;
 
-    public AgedBrieItem(String name, int sellIn, int quality) {
+    public NormalItem(String name, int sellIn, int quality) {
         this.name = name;
         this.sellIn = sellIn;
         this.quality = quality;
     }
+
     @Override
     public void update() {
         this.sellIn -= 1;
         if (this.sellIn >= 0) {
-            this.quality = Math.min(50, this.quality + 1);
+            this.quality = Math.max(0, this.quality - 1);
         } else {
-            this.quality = Math.min(50, this.quality + 2);
+            this.quality = Math.max(0, this.quality - 2);
         }
     }
 }
