+++ src/test/java/com/example/gildedrose/item/AgedBrieItemTest.java
@@ -1,7 +1,8 @@
-import gildedroseinventory.AgedBrieItem;
+package com.example.gildedrose.item;
+
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class AgedBrieItemTest {
     @Test
