+++ src/test/java/com/example/gildedrose/item/BackstagePassesItemTest.java
@@ -1,7 +1,8 @@
-import gildedroseinventory.BackstagePassesItem;
+package com.example.gildedrose.item;
+
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class BackstagePassesItemTest {
     @Test
