+++ src/test/java/com/example/gildedrose/item/ConjuredItemTest.java
@@ -1,7 +1,8 @@
-import gildedroseinventory.ConjuredItem;
+package com.example.gildedrose.item;
+
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class ConjuredItemTest {
     @Test
