+++ src/test/java/com/example/gildedrose/item/SulfurasItemTest.java
@@ -1,4 +1,6 @@
-import gildedroseinventory.SulfurasItem;
+package com.example.gildedrose.item;
+
+import com.example.gildedrose.item.SulfurasItem;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
